[refactor]: moved some extensions to views package
