+++ app/src/main/java/com/f0x1d/logfox/extensions/views/PreferenceExtensions.kt
@@ -1,4 +1,4 @@
-package com.f0x1d.logfox.extensions
+package com.f0x1d.logfox.extensions.views
 
 import android.view.LayoutInflater
 import android.view.WindowManager
