+++ app/src/main/java/com/f0x1d/logfox/ui/fragment/filters/EditFilterFragment.kt
@@ -13,7 +13,7 @@ import androidx.lifecycle.asLiveData
 import androidx.navigation.fragment.findNavController
 import com.f0x1d.logfox.R
 import com.f0x1d.logfox.databinding.FragmentEditFilterBinding
-import com.f0x1d.logfox.extensions.setClickListenerOn
+import com.f0x1d.logfox.extensions.views.setClickListenerOn
 import com.f0x1d.logfox.model.LogLevel
 import com.f0x1d.logfox.ui.fragment.base.BaseViewModelFragment
 import com.f0x1d.logfox.utils.event.Event
